# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ChainTrack Supply Chain Transparency Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List service requests
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [requested, accepted, in_progress, completed, cancelled]
        - name: role
          in: query
          schema:
            type: string
            enum: [customer, provider]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestList"

    post:
      operationId: createRequest
      tags: [requests]
      summary: Create a service request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestBody"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /requests/estimate:
    post:
      operationId: estimatePrice
      tags: [requests]
      summary: Get smart price estimate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateBody"
      responses:
        "200":
          description: Price estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceEstimate"
  /requests/available:
    get:
      operationId: listAvailableRequests
      tags: [requests]
      summary: List available requests for providers to accept
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Available requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestList"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get request by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests/{id}/status:
    patch:
      operationId: updateRequestStatus
      tags: [requests]
      summary: Update request status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /requests/{id}/accept:
    post:
      operationId: acceptRequest
      tags: [requests]
      summary: Provider accepts a request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
...
  /tracking/{requestId}:
    get:
      operationId: getTracking
      tags: [tracking]
      summary: Get tracking updates for a request
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tracking data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingData"

    post:
      operationId: addTrackingUpdate
      tags: [tracking]
      summary: Add a tracking update (provider only)
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackingUpdateBody"
      responses:
        "201":
          description: Update added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingUpdate"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, provider]
        phone:
          type: string
          nullable: true
        vehicleType:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        totalRatings:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, totalRatings, createdAt]
    CreateRequestBody:
      type: object
      properties:
        pickupLocation:
          type: string
        dropLocation:
          type: string
        pickupLat:
          type: number
          nullable: true
        pickupLng:
          type: number
          nullable: true
        dropLat:
          type: number
          nullable: true
        dropLng:
          type: number
          nullable: true
        description:
          type: string
        serviceType:
          type: string
          enum: [delivery, transport, logistics]
        offeredPrice:
          type: number
          nullable: true
        distanceKm:
          type: number
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
      required: [pickupLocation, dropLocation, description, serviceType]
    EstimateBody:
      type: object
      properties:
        pickupLocation:
          type: string
        dropLocation:
          type: string
        distanceKm:
          type: number
        serviceType:
          type: string
          enum: [delivery, transport, logistics]
      required: [pickupLocation, dropLocation, distanceKm, serviceType]
    PriceEstimate:
      type: object
      properties:
        baseFare:
          type: number
        distanceCharge:
          type: number
        serviceFee:
          type: number
        total:
          type: number
        breakdown:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              amount:
                type: number
            required: [label, amount]
        distanceKm:
          type: number
      required: [baseFare, distanceCharge, serviceFee, total, breakdown, distanceKm]
    ServiceRequest:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        providerId:
          type: integer
          nullable: true
        pickupLocation:
          type: string
        dropLocation:
          type: string
        pickupLat:
          type: number
          nullable: true
        pickupLng:
          type: number
          nullable: true
        dropLat:
          type: number
          nullable: true
        dropLng:
          type: number
          nullable: true
        description:
          type: string
        serviceType:
          type: string
          enum: [delivery, transport, logistics]
        status:
          type: string
          enum: [requested, accepted, in_progress, completed, cancelled]
        offeredPrice:
          type: number
          nullable: true
        agreedPrice:
          type: number
          nullable: true
        distanceKm:
          type: number
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        customer:
          $ref: "#/components/schemas/User"
        provider:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, customerId, pickupLocation, dropLocation, description, serviceType, status, createdAt, updatedAt, customer]
    RequestList:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRequest"
        total:
          type: integer
      required: [requests, total]
    UpdateStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [in_progress, completed, cancelled]
      required: [status]
...
    TrackingUpdate:
      type: object
      properties:
        id:
          type: integer
        requestId:
          type: integer
        status:
          type: string
        message:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, requestId, status, message, createdAt]
    TrackingData:
      type: object
      properties:
        requestId:
          type: integer
        currentStatus:
          type: string
        updates:
          type: array
          items:
            $ref: "#/components/schemas/TrackingUpdate"
        currentLat:
          type: number
          nullable: true
        currentLng:
          type: number
          nullable: true
      required: [requestId, currentStatus, updates]
    TrackingUpdateBody:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
      required: [status, message]
...