# lib/api-zod/src/generated/types/createRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChainTrack Supply Chain Transparency Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRequestBodyServiceType } from "./createRequestBodyServiceType";

export interface CreateRequestBody {
  pickupLocation: string;
  dropLocation: string;
  pickupLat?: number | null;
  pickupLng?: number | null;
  dropLat?: number | null;
  dropLng?: number | null;
  description: string;
  serviceType: CreateRequestBodyServiceType;
  offeredPrice?: number | null;
  distanceKm?: number | null;
  scheduledAt?: Date | null;
}