# lib/api-zod/src/generated/types/priceEstimate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChainTrack Supply Chain Transparency Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PriceEstimateBreakdownItem } from "./priceEstimateBreakdownItem";

export interface PriceEstimate {
  baseFare: number;
  distanceCharge: number;
  serviceFee: number;
  total: number;
  breakdown: PriceEstimateBreakdownItem[];
  distanceKm: number;
}